requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.4
boto3==1.42.59
botocore==1.42.59
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==5.2.7
django-bootstrap4==25.3
django-storages==1.14.6
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
factory_boy==3.0.1
Faker==38.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.2
